Use data generator
Model training and prediction now works off data generators instead
of in-memory data blocks. This keeps the training memory usage
O(1).

If all the data can fit in memory, it will be embedded once and kept
in memory during all of training. If it cannot, it will be embedded
in sequential blocks and each epoch will require its own parsing and
embedding.